Fix frontend Google Script run harness reuse
