+++ src/frontend/src/test/googleScriptRunHarness.ts
@@ -1,3 +1,5 @@
+import { createGoogleScriptRunApiHandlerMock as createGoogleScriptRunApiHandlerImplementation } from './google-script-run-harness-factory.js';
+
 export type GoogleScriptRunApiHandler = {
   withSuccessHandler: (
     handler: (response: unknown) => void
@@ -25,55 +27,8 @@ export type GoogleScriptRunApiHandlerCallbacks = Readonly<{
 export function createGoogleScriptRunApiHandlerMock(
   invokeRequest: (request: unknown, callbacks: GoogleScriptRunApiHandlerCallbacks) => void
 ): GoogleScriptRunApiHandler {
-  return createRunner();
-
-  /**
-   * Creates a runner snapshot with the currently registered handlers.
-   *
-   * @param {((response: unknown) => void) | undefined} successHandler The current success handler.
-   * @param {((error: unknown) => void) | undefined} failureHandler The current failure handler.
-   * @returns {GoogleScriptRunApiHandler} The runner snapshot.
-   */
-  function createRunner(
-    successHandler?: (response: unknown) => void,
-    failureHandler?: (error: unknown) => void
-  ): GoogleScriptRunApiHandler {
-    return {
-      withSuccessHandler(nextSuccessHandler: (response: unknown) => void) {
-        return createRunner(nextSuccessHandler, failureHandler);
-      },
-      withFailureHandler(nextFailureHandler: (error: unknown) => void) {
-        return createRunner(successHandler, nextFailureHandler);
-      },
-      apiHandler(request: unknown) {
-        invokeRequest(request, {
-          successHandler,
-          failureHandler,
-        });
-      },
-    };
-  }
+  return createGoogleScriptRunApiHandlerImplementation(invokeRequest) as GoogleScriptRunApiHandler;
 }
 
-export const googleScriptRunApiHandlerFactorySource = String.raw`
-function createGoogleScriptRunApiHandlerMock(invokeRequest) {
-  return createRunner();
-
-  function createRunner(successHandler, failureHandler) {
-    return {
-      withSuccessHandler(nextSuccessHandler) {
-        return createRunner(nextSuccessHandler, failureHandler);
-      },
-      withFailureHandler(nextFailureHandler) {
-        return createRunner(successHandler, nextFailureHandler);
-      },
-      apiHandler(request) {
-        invokeRequest(request, {
-          successHandler,
-          failureHandler,
-        });
-      },
-    };
-  }
-}
-`;
+export const googleScriptRunApiHandlerFactorySource =
+  createGoogleScriptRunApiHandlerImplementation.toString();
